backend: tidy core processor code. add tx hash and caller saving to db
